Move common routines to draw-common.c
Move the common routines twin_premultiply_alpha() and
_twin_apply_alpha() in src/draw.c to src/draw-common.c and put the
dedicated implementations into src/draw-builtin.c and
src/draw-pixman.c, respectively.
