+++ Makefile
@@ -44,6 +44,7 @@ libtwin.a_files-y = \
 	src/pattern.c \
 	src/spline.c \
 	src/work.c \
+	src/draw-common.c \
 	src/hull.c \
 	src/icon.c \
 	src/pixmap.c \
@@ -61,7 +62,7 @@ libtwin.a_files-$(CONFIG_LOGGING) += src/log.c
 libtwin.a_files-$(CONFIG_CURSOR) += src/cursor.c
 
 # Renderer
-libtwin.a_files-$(CONFIG_RENDERER_BUILTIN) += src/draw.c
+libtwin.a_files-$(CONFIG_RENDERER_BUILTIN) += src/draw-builtin.c
 libtwin.a_files-$(CONFIG_RENDERER_PIXMAN) += src/draw-pixman.c
 libtwin.a_cflags-$(CONFIG_RENDERER_PIXMAN) += $(shell pkg-config --cflags pixman-1)
 ifeq ($(CONFIG_RENDERER_PIXMAN), y)
